Change package's name
