+++ Task_1_1_3/src/main/java/ru/nsu/fit/labusov/expression/Add.java
@@ -1,42 +1,41 @@
-package nsu.fit.labusov;
+package ru.nsu.fit.labusov.expression;
 
 /**
- * Multiplication class.
+ * Addition class.
  */
-public class Mul extends Expression {
+public class Add extends Expression {
     private final Expression firstTerm;
     private final Expression secondTerm;
 
     /**
-     * Constructor mul method.
+     * Constructor add method.
      */
-    public Mul(Expression firstTerm, Expression secondTerm) {
+    public Add(Expression firstTerm, Expression secondTerm) {
         this.firstTerm = firstTerm;
         this.secondTerm = secondTerm;
     }
 
     @Override
     public String showExpression() {
-        return "(" + firstTerm.showExpression() + "*" + secondTerm.showExpression() + ")";
+        return "(" + firstTerm.showExpression() + "+" + secondTerm.showExpression() + ")";
     }
 
     @Override
     public void print() {
         System.out.print("(");
         firstTerm.print();
-        System.out.print("*");
+        System.out.print("+");
         secondTerm.print();
         System.out.print(")");
     }
 
     @Override
     public Expression derivative(String string) {
-        return new Add(new Mul(firstTerm.derivative(string), secondTerm),
-                new Mul(firstTerm, secondTerm.derivative(string)));
+        return new Add(firstTerm.derivative(string), secondTerm.derivative(string));
     }
 
     @Override
     public int eval(String string) {
-        return firstTerm.eval(string) * secondTerm.eval(string);
+        return firstTerm.eval(string) + secondTerm.eval(string);
     }
 }
