+++ Task_1_1_3/src/main/java/ru/nsu/fit/labusov/expression/Div.java
@@ -1,4 +1,4 @@
-package nsu.fit.labusov;
+package ru.nsu.fit.labusov.expression;
 
 /**
  * Division class.
