+++ Task_1_1_3/src/main/java/ru/nsu/fit/labusov/expression/Mul.java
@@ -1,41 +1,42 @@
-package nsu.fit.labusov;
+package ru.nsu.fit.labusov.expression;
 
 /**
- * Addition class.
+ * Multiplication class.
  */
-public class Add extends Expression {
+public class Mul extends Expression {
     private final Expression firstTerm;
     private final Expression secondTerm;
 
     /**
-     * Constructor add method.
+     * Constructor mul method.
      */
-    public Add(Expression firstTerm, Expression secondTerm) {
+    public Mul(Expression firstTerm, Expression secondTerm) {
         this.firstTerm = firstTerm;
         this.secondTerm = secondTerm;
     }
 
     @Override
     public String showExpression() {
-        return "(" + firstTerm.showExpression() + "+" + secondTerm.showExpression() + ")";
+        return "(" + firstTerm.showExpression() + "*" + secondTerm.showExpression() + ")";
     }
 
     @Override
     public void print() {
         System.out.print("(");
         firstTerm.print();
-        System.out.print("+");
+        System.out.print("*");
         secondTerm.print();
         System.out.print(")");
     }
 
     @Override
     public Expression derivative(String string) {
-        return new Add(firstTerm.derivative(string), secondTerm.derivative(string));
+        return new Add(new Mul(firstTerm.derivative(string), secondTerm),
+                new Mul(firstTerm, secondTerm.derivative(string)));
     }
 
     @Override
     public int eval(String string) {
-        return firstTerm.eval(string) + secondTerm.eval(string);
+        return firstTerm.eval(string) * secondTerm.eval(string);
     }
 }
